requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
SQLAlchemy==2.0.29
Werkzeug==3.0.2
Jinja2==3.1.3
python-dotenv==1.0.1
openai==1.65.3
requests==2.31.0
email-validator==2.1.1
pytest==8.0.0
